requirements.txt
pyaudio  
wave  
numpy
requests
python-dotenv
pydub
